requirements.txt
python-telegram-bot==13.15
beautifulsoup4
requests
flask           
gunicorn        